+++ includes/Setup/Editor.php
@@ -9,7 +9,8 @@
  * @package    Zooey
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.1
  */
 
 namespace WebManDesign\Zooey\Setup;
@@ -27,25 +28,34 @@ class Editor implements Component_Interface {
 	/**
 	 * Color theme options which CSS variables are used for setting up gradients in `theme.json`.
 	 *
-	 * Replace these color CSS variables with actual color values
+	 * Replaces these color CSS variables with actual color values
 	 * for block editor gradient control.
 	 *
-	 * IMPORTANT:
-	 * No need to add theme palette colors anymore since WordPress 6.5
-	 * supports CSS variables in gradient control.
-	 * Only really special (non-palette, automatically calculated)
-	 * colors are required now. (See `self::init()` below.)
+	 * NOTE:
+	 * WP6.5+ displays gradient control with correct colors even when
+	 * CSS variables are used, but the actual edit portion of gradient
+	 * control works only with absolute color values, not with CSS vars.
+	 * So, even in WP6.5+ we need to set this variable.
 	 *
 	 * @see  WebManDesign\Zooey\Customize\Options::set()
 	 * @see  theme.json/settings.color.gradients
 	 *
-	 * @since   1.0.0
-	 * @access  public
-	 * @var     array
+	 * @since    1.0.0
+	 * @version  1.0.1
+	 * @access   public
+	 * @var      array
 	 */
 	public static $color_options_special = array(
-		'color_secondary_semitransparent',
+		'color_base',
+		'color_base_alt',
+		'color_contrast',
+		'color_contrast_alt',
+		'color_primary',
+		'color_primary_mixed',
 		'color_primary_semitransparent',
+		'color_secondary',
+		'color_secondary_mixed',
+		'color_secondary_semitransparent',
 	);
 
 	/**
@@ -69,33 +79,14 @@ class Editor implements Component_Interface {
 	/**
 	 * Initialization.
 	 *
-	 * @since  1.0.0
+	 * @since    1.0.0
+	 * @version  1.0.1
 	 *
 	 * @return  void
 	 */
 	public static function init() {
 
 		// Processing
-
-			/**
-			 * WordPress older than 6.5 gradient control does not work
-			 * with CSS variables, so we need to add all theme palette
-			 * colors which CSS variables we used to set our gradients
-			 * in `theme.json`.
-			 */
-			if ( version_compare( $GLOBALS['wp_version'], '6.5', '<' ) ) {
-				array_push( self::$color_options_special,
-					// Order is important here!
-					'color_base_alt',
-					'color_base',
-					'color_contrast_alt',
-					'color_contrast',
-					'color_primary_mixed',
-					'color_primary',
-					'color_secondary_mixed',
-					'color_secondary'
-				);
-			}
 
 			// Filters