+++ changelog.md
@@ -1,5 +1,38 @@
 # Zooey Changelog
 
+## 1.0.1, 20240504
+
+### Added
+- New intro pattern
+
+### Updated
+- Not hiding global styles elements color controls so user can still set "Link" color (other controls are obsolete)
+- Improving patterns for demo website
+- Removing WP.org not-allowed links
+- Localization
+- Screenshot
+
+### Fixed
+- Editor gradient control colors
+- Pattern titles
+
+### File updates
+	changelog.md
+	style.css
+	includes/Assets/Editor.php
+	includes/Content/Block_Pattern.php
+	includes/Setup/Editor.php
+	languages/sk_SK.*
+	languages/zooey.pot
+	patterns/intro/intro-01.php
+	patterns/intro/intro-05.php
+	patterns/intro/intro-06.php
+	patterns/page/home-1.php
+	patterns/site/footer.php
+	patterns/site/header.php
+	patterns/site/sidebar.php
+
+
 ## 1.0.0, 20240504
 
 - Initial release.